apps/plant-api/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.9
django-3-jet==1.0.8
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-resized==1.0.2
django-tinymce==3.6.1
django-tinymce4-lite==1.8.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environ==1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsmin==3.0.1
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pyenchant==3.2.2
PyJWT==2.7.0
pyparsing==3.0.9
python-environ==0.4.54
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
